Clean up: extract shared collapse helper and dedupe mutation checks
