+++ src/yaml.rs
@@ -188,6 +188,87 @@ pub(crate) fn detach_empty_collection_placeholder_newline(
     }
 }
 
+/// If `collection` is an empty block MAPPING or SEQUENCE sitting
+/// inside a MAPPING_ENTRY's VALUE, replace the `NEWLINE INDENT
+/// COLLECTION` placeholder with an inline flow-empty
+/// (` {}` or ` []`) so the entry renders as `key: {}` / `key: []`
+/// instead of an unterminated `key:\n    ` shape.
+///
+/// Skips flow collections (`{}` / `[]` already render correctly) and
+/// VALUEs carrying anything beyond decoration around the collection
+/// (comments, sibling scalars, anchors, tags).
+pub(crate) fn collapse_empty_child_collection_in_parent(collection: &SyntaxNode) {
+    let (open_kind, close_kind, open_txt, close_txt) = match collection.kind() {
+        SyntaxKind::MAPPING => (SyntaxKind::LEFT_BRACE, SyntaxKind::RIGHT_BRACE, "{", "}"),
+        SyntaxKind::SEQUENCE => (
+            SyntaxKind::LEFT_BRACKET,
+            SyntaxKind::RIGHT_BRACKET,
+            "[",
+            "]",
+        ),
+        _ => return,
+    };
+
+    // Flow collections already render correctly when empty.
+    let is_flow = collection
+        .children_with_tokens()
+        .any(|c| c.as_token().is_some_and(|t| t.kind() == open_kind));
+    if is_flow {
+        return;
+    }
+
+    let Some(value_node) = collection.parent() else {
+        return;
+    };
+    if value_node.kind() != SyntaxKind::VALUE {
+        return;
+    }
+    let Some(entry_node) = value_node.parent() else {
+        return;
+    };
+    if entry_node.kind() != SyntaxKind::MAPPING_ENTRY {
+        return;
+    }
+
+    // Bail if the VALUE carries anything beyond decoration.
+    let has_other = value_node.children_with_tokens().any(|el| match el {
+        rowan::NodeOrToken::Node(n) => &n != collection,
+        rowan::NodeOrToken::Token(t) => !matches!(
+            t.kind(),
+            SyntaxKind::NEWLINE | SyntaxKind::INDENT | SyntaxKind::WHITESPACE
+        ),
+    });
+    if has_other {
+        return;
+    }
+
+    let mut builder = rowan::GreenNodeBuilder::new();
+    builder.start_node(SyntaxKind::VALUE.into());
+    builder.token(SyntaxKind::WHITESPACE.into(), " ");
+    builder.start_node(collection.kind().into());
+    builder.token(open_kind.into(), open_txt);
+    builder.token(close_kind.into(), close_txt);
+    builder.finish_node();
+    builder.finish_node();
+    let new_value = SyntaxNode::new_root_mut(builder.finish());
+
+    let Some(value_idx) = entry_node
+        .children_with_tokens()
+        .position(|c| c.as_node() == Some(&value_node))
+    else {
+        return;
+    };
+    entry_node.splice_children(value_idx..(value_idx + 1), vec![new_value.into()]);
+
+    // The fresh VALUE doesn't carry a NEWLINE; without one the next
+    // sibling entry would glue onto our tail.
+    if !ends_with_newline(&entry_node) {
+        let nl = crate::nodes::fresh_token(SyntaxKind::NEWLINE, "\n");
+        let end = entry_node.children_with_tokens().count();
+        entry_node.splice_children(end..end, vec![nl.into()]);
+    }
+}
+
 /// Create a newline token and add it to the elements vector
 pub(crate) fn add_newline_token(
     elements: &mut Vec<rowan::NodeOrToken<rowan::SyntaxNode<Lang>, rowan::SyntaxToken<Lang>>>,
