+++ tests/proptest_invariants.rs
@@ -8,7 +8,10 @@ use proptest::prelude::*;
 use rowan::ast::AstNode;
 use std::str::FromStr;
 use yaml_edit::path::YamlPath;
-use yaml_edit::{debug, Document, Mapping, Sequence};
+use yaml_edit::{debug, Document};
+
+mod common;
+use common::mutation_checks::*;
 
 /// A random-but-plausible mutation to apply.
 ///
@@ -174,443 +177,6 @@ fn seed_strat() -> impl Strategy<Value = &'static str> {
     ]
 }
 
-// Sequence-mutation post-conditions. Kept in sync with the fuzz target's
-// versions -- see fuzz/fuzz_targets/mutation_invariants.rs for the full
-// rationale on each check.
-
-fn reparse_seq(doc: &Document, key: &str, op: &str) -> Sequence {
-    let text = doc.to_string();
-    let reparsed = Document::from_str(&text)
-        .unwrap_or_else(|e| panic!("{op}: re-parse failed ({e}), text: {text:?}"));
-    reparsed
-        .as_mapping()
-        .and_then(|m| m.get_sequence(key))
-        .unwrap_or_else(|| {
-            panic!("{op}: re-parsed doc has no sequence at key {key:?}, text: {text:?}")
-        })
-}
-
-fn scalar_at(seq: &Sequence, index: usize) -> Option<String> {
-    seq.get(index)
-        .as_ref()
-        .and_then(|n| n.as_scalar().map(|s| s.as_string()))
-}
-
-fn assert_seq_push_stuck(seq: &Sequence, before: usize, v: &str, doc: &Document, key: &str) {
-    let op = format!("SeqPush({v:?})");
-    let after = seq.len();
-    if after != before + 1 {
-        panic!(
-            "{op}: len {before} -> {after} (expected {}), text: {:?}",
-            before + 1,
-            doc.to_string()
-        );
-    }
-    let last = scalar_at(seq, after - 1);
-    if last.as_deref() != Some(v) {
-        panic!("{op}: last item = {last:?}, text: {:?}", doc.to_string());
-    }
-    let reparsed = reparse_seq(doc, key, &op);
-    let r_len = reparsed.len();
-    let r_last = scalar_at(&reparsed, r_len.saturating_sub(1));
-    if r_len != after || r_last.as_deref() != Some(v) {
-        panic!(
-            "{op}: reparse drift: len {after} vs {r_len}, last {:?} vs {r_last:?}, text: {:?}",
-            Some(v),
-            doc.to_string()
-        );
-    }
-}
-
-fn assert_seq_pop_stuck(seq: &Sequence, before: usize, popped: bool, doc: &Document, key: &str) {
-    let op = "SeqPop";
-    let after = seq.len();
-    let expected = if popped { before - 1 } else { before };
-    if after != expected {
-        panic!(
-            "{op}: popped={popped} len {before} -> {after} (expected {expected}), text: {:?}",
-            doc.to_string()
-        );
-    }
-    if popped && before == 0 {
-        panic!(
-            "{op}: reported pop of an empty sequence, text: {:?}",
-            doc.to_string()
-        );
-    }
-    if !popped && before > 0 {
-        panic!(
-            "{op}: refused to pop a non-empty sequence, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_seq(doc, key, op);
-    if reparsed.len() != after {
-        panic!(
-            "{op}: reparse drift: len {after} vs {}, text: {:?}",
-            reparsed.len(),
-            doc.to_string()
-        );
-    }
-}
-
-fn assert_seq_insert_stuck(
-    seq: &Sequence,
-    before: usize,
-    i: usize,
-    v: &str,
-    doc: &Document,
-    key: &str,
-) {
-    let op = format!("SeqInsert({i}, {v:?})");
-    let after = seq.len();
-    if after != before + 1 {
-        panic!(
-            "{op}: len {before} -> {after} (expected {}), text: {:?}",
-            before + 1,
-            doc.to_string()
-        );
-    }
-    let effective = i.min(before);
-    let at = scalar_at(seq, effective);
-    if at.as_deref() != Some(v) {
-        panic!(
-            "{op}: item at effective index {effective} = {at:?}, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_seq(doc, key, &op);
-    let r_at = scalar_at(&reparsed, effective);
-    if reparsed.len() != after || r_at.as_deref() != Some(v) {
-        panic!(
-            "{op}: reparse drift: len {after} vs {}, at {effective} {:?} vs {r_at:?}, text: {:?}",
-            reparsed.len(),
-            Some(v),
-            doc.to_string()
-        );
-    }
-}
-
-fn assert_seq_set_stuck(
-    seq: &Sequence,
-    before: usize,
-    i: usize,
-    v: &str,
-    ok: bool,
-    doc: &Document,
-    key: &str,
-) {
-    let op = format!("SeqSet({i}, {v:?})");
-    let after = seq.len();
-    if after != before {
-        panic!(
-            "{op}: ok={ok} len {before} -> {after} (expected unchanged), text: {:?}",
-            doc.to_string()
-        );
-    }
-    if ok != (i < before) {
-        panic!(
-            "{op}: ok={ok} but i={i} vs before={before}, text: {:?}",
-            doc.to_string()
-        );
-    }
-    if ok {
-        let at = scalar_at(seq, i);
-        if at.as_deref() != Some(v) {
-            panic!("{op}: item at {i} = {at:?}, text: {:?}", doc.to_string());
-        }
-        let reparsed = reparse_seq(doc, key, &op);
-        let r_at = scalar_at(&reparsed, i);
-        if reparsed.len() != after || r_at.as_deref() != Some(v) {
-            panic!(
-                "{op}: reparse drift: len {after} vs {}, at {i} {:?} vs {r_at:?}, text: {:?}",
-                reparsed.len(),
-                Some(v),
-                doc.to_string()
-            );
-        }
-    }
-}
-
-fn assert_seq_remove_stuck(
-    seq: &Sequence,
-    before: usize,
-    i: usize,
-    removed: bool,
-    doc: &Document,
-    key: &str,
-) {
-    let op = format!("SeqRemove({i})");
-    let after = seq.len();
-    let expected = if removed { before - 1 } else { before };
-    if after != expected {
-        panic!(
-            "{op}: removed={removed} len {before} -> {after} (expected {expected}), text: {:?}",
-            doc.to_string()
-        );
-    }
-    if removed != (i < before) {
-        panic!(
-            "{op}: removed={removed} but i={i} vs before={before}, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_seq(doc, key, &op);
-    if reparsed.len() != after {
-        panic!(
-            "{op}: reparse drift: len {after} vs {}, text: {:?}",
-            reparsed.len(),
-            doc.to_string()
-        );
-    }
-}
-
-fn assert_seq_clear_stuck(seq: &Sequence, doc: &Document, key: &str) {
-    let op = "SeqClear";
-    if !seq.is_empty() {
-        panic!(
-            "{op}: seq not empty after clear (len={}), text: {:?}",
-            seq.len(),
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_seq(doc, key, op);
-    if !reparsed.is_empty() {
-        panic!(
-            "{op}: reparse drift: reparsed len {} (expected 0), text: {:?}",
-            reparsed.len(),
-            doc.to_string()
-        );
-    }
-}
-
-// Mapping-mutation post-conditions.
-
-fn reparse_mapping(doc: &Document, op: &str) -> Mapping {
-    let text = doc.to_string();
-    let reparsed = Document::from_str(&text)
-        .unwrap_or_else(|e| panic!("{op}: re-parse failed ({e}), text: {text:?}"));
-    reparsed
-        .as_mapping()
-        .unwrap_or_else(|| panic!("{op}: re-parsed doc is not a mapping, text: {text:?}"))
-}
-
-fn scalar_value(m: &Mapping, key: &str) -> Option<String> {
-    m.get(key)
-        .as_ref()
-        .and_then(|n| n.as_scalar().map(|s| s.as_string()))
-}
-
-/// True if `mapping`'s enclosing container is a MAPPING_ENTRY (nested
-/// under a key). Top-level mappings live directly under the DOCUMENT
-/// node.
-fn is_nested_mapping(mapping: &Mapping) -> bool {
-    use yaml_edit::SyntaxKind;
-    mapping
-        .syntax()
-        .parent()
-        .filter(|p| p.kind() == SyntaxKind::VALUE)
-        .and_then(|v| v.parent())
-        .is_some_and(|e| e.kind() == SyntaxKind::MAPPING_ENTRY)
-}
-
-/// Verify that `mapping.set(k, v)` for a string value landed: the key
-/// exists in memory with scalar value `v`, and the same holds after
-/// re-parse.
-fn assert_mapping_set_stuck(mapping: &Mapping, k: &str, v: &str, doc: &Document) {
-    let op = format!("MappingSet({k:?}, {v:?})");
-    if !mapping.contains_key(k) {
-        panic!("{op}: key missing after set, text: {:?}", doc.to_string());
-    }
-    let got = scalar_value(mapping, k);
-    if got.as_deref() != Some(v) {
-        panic!(
-            "{op}: value at {k:?} = {got:?}, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_mapping(doc, &op);
-    if !reparsed.contains_key(k) {
-        panic!(
-            "{op}: reparse drift: key {k:?} missing, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let r_got = scalar_value(&reparsed, k);
-    if r_got.as_deref() != Some(v) {
-        panic!(
-            "{op}: reparse drift: value at {k:?} = {r_got:?} (expected {v:?}), text: {:?}",
-            doc.to_string()
-        );
-    }
-}
-
-/// Verify that `mapping.remove(k)` did what it claimed.
-fn assert_mapping_remove_stuck(mapping: &Mapping, k: &str, existed: bool, doc: &Document) {
-    let op = format!("MappingRemove({k:?})");
-    if existed && mapping.contains_key(k) {
-        panic!(
-            "{op}: reported removed but key still present, text: {:?}",
-            doc.to_string()
-        );
-    }
-    if !existed && mapping.contains_key(k) {
-        // remove returned None but key is present -- that's a
-        // pre-existing state, not our concern.
-        return;
-    }
-    // An emptied top-level mapping renders as `""` which re-parses as
-    // no mapping (see the note in Mapping::remove_entry_at). That's
-    // an accepted intermediate state, not a bug: nothing was lost
-    // because there was nothing left to save. Skip the re-parse check
-    // in that case.
-    if mapping.is_empty() && !is_nested_mapping(mapping) {
-        return;
-    }
-    let reparsed = reparse_mapping(doc, &op);
-    if existed && reparsed.contains_key(k) {
-        panic!(
-            "{op}: reparse drift: key {k:?} came back, text: {:?}",
-            doc.to_string()
-        );
-    }
-}
-
-/// Verify that `mapping.rename_key(old, new)` returned truthfully.
-/// `rename_key` renames only the first occurrence of `old`, so with
-/// duplicate keys `old` may legitimately still be present after a
-/// successful rename; only the presence of `new` is a hard requirement.
-fn assert_mapping_rename_stuck(
-    mapping: &Mapping,
-    _old: &str,
-    new: &str,
-    renamed: bool,
-    doc: &Document,
-) {
-    let op = format!("MappingRename(-> {new:?})");
-    if renamed {
-        if !mapping.contains_key(new) {
-            panic!(
-                "{op}: reported rename but new key missing, text: {:?}",
-                doc.to_string()
-            );
-        }
-        let reparsed = reparse_mapping(doc, &op);
-        if !reparsed.contains_key(new) {
-            panic!(
-                "{op}: reparse drift: new key {new:?} missing, text: {:?}",
-                doc.to_string()
-            );
-        }
-    }
-}
-
-/// Verify that `mapping.clear()` emptied the mapping in memory and
-/// after re-parse.
-fn assert_mapping_clear_stuck(mapping: &Mapping, doc: &Document) {
-    let op = "MappingClear";
-    if !mapping.is_empty() {
-        panic!("{op}: not empty in memory, text: {:?}", doc.to_string());
-    }
-    let reparsed = reparse_mapping(doc, op);
-    if !reparsed.is_empty() {
-        panic!("{op}: not empty after reparse, text: {:?}", doc.to_string());
-    }
-}
-
-/// Verify that a mapping-insert op (insert_after/before/at_index) put
-/// the key in the mapping with the requested scalar value, and that
-/// this survives re-parse.
-fn assert_mapping_insert_stuck(mapping: &Mapping, k: &str, v: &str, doc: &Document, op_name: &str) {
-    let op = format!("{op_name}({k:?}, {v:?})");
-    if !mapping.contains_key(k) {
-        panic!(
-            "{op}: key missing after insert, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let got = scalar_value(mapping, k);
-    if got.as_deref() != Some(v) {
-        panic!(
-            "{op}: value at {k:?} = {got:?}, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let reparsed = reparse_mapping(doc, &op);
-    if !reparsed.contains_key(k) {
-        panic!(
-            "{op}: reparse drift: key {k:?} missing, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let r_got = scalar_value(&reparsed, k);
-    if r_got.as_deref() != Some(v) {
-        panic!(
-            "{op}: reparse drift: value at {k:?} = {r_got:?}, text: {:?}",
-            doc.to_string()
-        );
-    }
-}
-
-/// Verify that `doc.set_path(path, v)` landed `v` at `path` -- both
-/// in-memory and after re-parse. set_path is a no-op when the doc has
-/// no root mapping, so an empty-doc call is not a failure.
-fn assert_set_path_stuck(doc: &Document, path: &str, v: &str) {
-    let op = format!("SetPath({path:?}, {v:?})");
-    // If the doc has no root mapping, set_path silently returned.
-    if doc.as_mapping().is_none() {
-        return;
-    }
-    let got = doc
-        .get_path(path)
-        .as_ref()
-        .and_then(|n| n.as_scalar().map(|s| s.as_string()));
-    if got.as_deref() != Some(v) {
-        panic!(
-            "{op}: get_path returned {got:?} in memory, text: {:?}",
-            doc.to_string()
-        );
-    }
-    let text = doc.to_string();
-    let Ok(reparsed) = Document::from_str(&text) else {
-        panic!("{op}: re-parse failed, text: {text:?}");
-    };
-    let r_got = reparsed
-        .get_path(path)
-        .as_ref()
-        .and_then(|n| n.as_scalar().map(|s| s.as_string()));
-    if r_got.as_deref() != Some(v) {
-        panic!("{op}: reparse drift: get_path returned {r_got:?}, text: {text:?}");
-    }
-}
-
-/// Verify that `doc.remove_path(path)` did what it claimed and that
-/// the path is gone after re-parse.
-fn assert_remove_path_stuck(doc: &Document, path: &str, removed: bool) {
-    let op = format!("RemovePath({path:?})");
-    if doc.as_mapping().is_none() {
-        return;
-    }
-    if removed && doc.get_path(path).is_some() {
-        panic!(
-            "{op}: reported removed but path still present, text: {:?}",
-            doc.to_string()
-        );
-    }
-    // Only require re-parse to hide the path when remove reported
-    // success; a `false` return may leave a pre-existing state.
-    if !removed {
-        return;
-    }
-    let text = doc.to_string();
-    let Ok(reparsed) = Document::from_str(&text) else {
-        panic!("{op}: re-parse failed, text: {text:?}");
-    };
-    if reparsed.get_path(path).is_some() {
-        panic!("{op}: reparse drift: path came back, text: {text:?}");
-    }
-}
-
 fn apply(doc: &Document, op: &Op) {
     let Some(mapping) = doc.as_mapping() else {
         return;
@@ -641,7 +207,7 @@ fn apply(doc: &Document, op: &Op) {
         }
         Op::Rename(a, b) => {
             let renamed = mapping.rename_key(a.as_str(), b.as_str());
-            assert_mapping_rename_stuck(&mapping, a.as_str(), b.as_str(), renamed, doc);
+            assert_mapping_rename_stuck(&mapping, b.as_str(), renamed, doc);
         }
         Op::Clear => {
             mapping.clear();
